+++ package.json
@@ -2,7 +2,7 @@
   "name": "sunblock",
   "version": "1.0.0",
   "description": "",
-  "main": "sunblock.js",
+  "main": "src/sunblock.js",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
